requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attr==0.3.1
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
cycler==0.11.0
Cython==0.29.36
deep-sort-realtime==1.3.2
deepface==0.0.79
defusedxml==0.7.1
Django==3.2.19
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
expiringdict==1.2.2
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
google-api-core==2.11.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.56.0
grpcio-status==1.48.2
gunicorn==20.1.0
h5py==3.9.0
helper==2.5.0
htmlmin==0.1.12
huggingface-hub==0.16.4
idna==3.4
ijson==3.2.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jsonschema==3.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
label-studio==1.8.0
label-studio-converter==0.0.53
label-studio-tools==0.0.2
lapx==0.5.2.post1
launchdarkly-server-sdk==7.5.0
libclang==16.0.0
lockfile==0.12.2
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.8.11
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
nltk==3.6.7
numpy==1.25.1
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.1
pandas==1.5.3
Pillow==9.3.0
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
redis==3.5.3
regex==2023.6.3
requests==2.28.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
safetensors==0.3.1
scipy==1.11.1
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.28.0
simplejson==3.19.1
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
timm==0.9.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
transliterate==1.10.2
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
ujson==5.8.0
ultralytics==8.0.136
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
xmljson==0.2.0
zipp==3.15.0
